+++ app/client/cli/debug.go
@@ -1,10 +1,15 @@
 package cli
 
 import (
+	"fmt"
+	"log"
 	"os"
+	"os/exec"
+	"time"
 
 	"github.com/manifoldco/promptui"
 	"github.com/spf13/cobra"
+	"golang.org/x/exp/slices"
 	"google.golang.org/protobuf/types/known/anypb"
 
 	"github.com/pokt-network/pocket/app/client/cli/helpers"
@@ -34,43 +39,123 @@ var items = []string{
 }
 
 func init() {
+	dbg := newDebugCommand()
+	dbg.AddCommand(newDebugSubCommands()...)
+	rootCmd.AddCommand(dbg)
+
 	dbgUI := newDebugUICommand()
-	dbgUI.AddCommand(newDebugUISubCommands()...)
 	rootCmd.AddCommand(dbgUI)
 }
 
-// newDebugUISubCommands builds out the list of debug subcommands by matching the
-// handleSelect dispatch to the appropriate command.
-// * To add a debug subcommand, you must add it to the `items` array and then
-// write a function handler to match for it in `handleSelect`.
-func newDebugUISubCommands() []*cobra.Command {
-	commands := make([]*cobra.Command, len(items))
-	for idx, promptItem := range items {
-		commands[idx] = &cobra.Command{
-			Use:               promptItem,
-			PersistentPreRunE: helpers.P2PDependenciesPreRunE,
-			Run: func(cmd *cobra.Command, _ []string) {
-				handleSelect(cmd, cmd.Use)
-			},
-			ValidArgs: items,
-		}
+// newDebugCommand returns the cobra CLI for the Debug command.
+func newDebugCommand() *cobra.Command {
+	return &cobra.Command{
+		Use:     "Debug",
+		Aliases: []string{"d"},
+		Short:   "Debug utility for rapid development",
+		Long:    "Debug utility to send fire-and-forget messages to the network for development purposes",
+		Args:    cobra.MaximumNArgs(1),
 	}
-	return commands
+}
+
+// newDebugSubCommands is a list of commands that can be "fired & forgotten" (no selection necessary)
+func newDebugSubCommands() []*cobra.Command {
+	cmds := []*cobra.Command{
+		{
+			Use:               "PrintNodeState",
+			Aliases:           []string{"print", "state"},
+			Short:             "Prints the node state",
+			Long:              "Sends a message to all visible nodes to log the current state of their consensus",
+			Args:              cobra.ExactArgs(0),
+			PersistentPreRunE: helpers.P2PDependenciesPreRunE,
+			Run: func(cmd *cobra.Command, args []string) {
+				runWithSleep(func() {
+					handleSelect(cmd, PromptPrintNodeState)
+				})
+			},
+		},
+		{
+			Use:               "ResetToGenesis",
+			Aliases:           []string{"reset", "genesis"},
+			Short:             "Reset to genesis",
+			Long:              "Broadcast a message to all visible nodes to reset the state to genesis",
+			Args:              cobra.ExactArgs(0),
+			PersistentPreRunE: helpers.P2PDependenciesPreRunE,
+			Run: func(cmd *cobra.Command, args []string) {
+				runWithSleep(func() {
+					handleSelect(cmd, PromptResetToGenesis)
+				})
+			},
+		},
+		{
+			Use:               "TriggerView",
+			Aliases:           []string{"next", "trigger", "view"},
+			Short:             "Trigger the next view in consensus",
+			Long:              "Sends a message to all visible nodes on the network to start the next view (height/step/round) in consensus",
+			Args:              cobra.ExactArgs(0),
+			PersistentPreRunE: helpers.P2PDependenciesPreRunE,
+			Run: func(cmd *cobra.Command, args []string) {
+				runWithSleep(func() {
+					handleSelect(cmd, PromptTriggerNextView)
+				})
+			},
+		},
+		{
+			Use:               "TogglePacemakerMode",
+			Aliases:           []string{"toggle", "pcm"},
+			Short:             "Toggle the pacemaker",
+			Long:              "Toggle the consensus pacemaker either on or off so the chain progresses on its own or loses liveness",
+			Args:              cobra.ExactArgs(0),
+			PersistentPreRunE: helpers.P2PDependenciesPreRunE,
+			Run: func(cmd *cobra.Command, args []string) {
+				runWithSleep(func() {
+					handleSelect(cmd, PromptTogglePacemakerMode)
+				})
+			},
+		},
+		{
+			Use:               "ScaleActor",
+			Aliases:           []string{"scale"},
+			Short:             "Scales the number of actors up or down",
+			Long:              "Scales the type of actor specified to the number provided",
+			Args:              cobra.ExactArgs(2),
+			PersistentPreRunE: helpers.P2PDependenciesPreRunE,
+			Run: func(cmd *cobra.Command, args []string) {
+				actor := args[0]
+				numActors := args[1]
+				validActors := []string{"fishermen", "full_nodes", "servicers", "validators"}
+				if !slices.Contains(validActors, actor) {
+					logger.Global.Fatal().Msg("Invalid actor type provided")
+				}
+				sedCmd := exec.Command("sed", "-i", fmt.Sprintf("/%s:/,/count:/ s/count: [0-9]*/count: %s/", actor, numActors), "/usr/local/localnet_config.yaml")
+				err := sedCmd.Run()
+				if err != nil {
+					log.Fatal(err)
+				}
+			},
+		},
+	}
+	return cmds
 }
 
 // newDebugUICommand returns the cobra CLI for the Debug UI interface.
 func newDebugUICommand() *cobra.Command {
 	return &cobra.Command{
-		Aliases:           []string{"dui"},
 		Use:               "DebugUI",
-		Short:             "Debug selection ui for rapid development",
+		Aliases:           []string{"dui"},
+		Short:             "Debug utility with an interactive UI for development purposes",
+		Long:              "Opens a shell-driven selection UI to view and select from a list of debug actions for development purposes",
 		Args:              cobra.MaximumNArgs(0),
 		PersistentPreRunE: helpers.P2PDependenciesPreRunE,
-		RunE:              runDebug,
+		RunE:              selectDebugCommand,
 	}
 }
 
-func runDebug(cmd *cobra.Command, _ []string) (err error) {
+// selectDebugCommand builds out the list of debug subcommands by matching the
+// handleSelect dispatch to the appropriate command.
+//   - To add a debug subcommand, you must add it to the `items` array and then
+//     write a function handler to match for it in `handleSelect`.
+func selectDebugCommand(cmd *cobra.Command, _ []string) error {
 	for {
 		if selection, err := promptGetInput(); err == nil {
 			handleSelect(cmd, selection)
@@ -158,7 +243,17 @@ func handleSelect(cmd *cobra.Command, selection string) {
 	}
 }
 
-// Broadcast to the entire network.
+// HACK: Because of how the p2p module works, we need to surround it with sleep both BEFORE and AFTER the task.
+// - Starting the task too early after the debug client initializes results in a lack of visibility of the nodes in the network
+// - Ending the task too early before the debug client completes its task results in a lack of propagation of the message or retrieval of the result
+// TECHDEBT: There is likely an event based solution to this but it would require a lot more refactoring of the p2p module.
+func runWithSleep(task func()) {
+	time.Sleep(1000 * time.Millisecond)
+	task()
+	time.Sleep(1000 * time.Millisecond)
+}
+
+// broadcastDebugMessage broadcasts the debug message to the entire visible network.
 func broadcastDebugMessage(cmd *cobra.Command, debugMsg *messaging.DebugMessage) {
 	anyProto, err := anypb.New(debugMsg)
 	if err != nil {
@@ -174,7 +269,7 @@ func broadcastDebugMessage(cmd *cobra.Command, debugMsg *messaging.DebugMessage)
 	}
 }
 
-// Send to just a single (i.e. first) validator in the set
+// sendDebugMessage sends the debug message to just a single (i.e. first) node visible
 func sendDebugMessage(cmd *cobra.Command, debugMsg *messaging.DebugMessage) {
 	anyProto, err := anypb.New(debugMsg)
 	if err != nil {